Refactor dev panel
