+++ src/components/DevModePanel/useDevModeKeyTrigger.ts
@@ -1,23 +1,22 @@
 import { useEffect } from "react";
 
-type DevModeKeyTriggerCallback = (
-  err: Error | null,
-  { shift }: { shift: boolean },
-) => void;
+type DevModeKeyTriggerCallback = ({ shift }: { shift: boolean }) => void;
 
-export const useDevModeKeyTrigger = (callback: DevModeKeyTriggerCallback) => {
+export const useDevModeKeyTrigger = (
+  callbackHandler: DevModeKeyTriggerCallback,
+) => {
   useEffect(() => {
     const handler = (event: KeyboardEvent) => {
       if (event.shiftKey && event.metaKey && event.code === "Quote") {
-        callback(null, { shift: true });
+        callbackHandler({ shift: true });
       } else if (event.metaKey && event.code === "Quote") {
-        callback(null, { shift: false });
+        callbackHandler({ shift: false });
       } else if (event.ctrlKey && event.code === "Quote") {
-        callback(null, { shift: false });
+        callbackHandler({ shift: false });
       }
     };
 
     document.addEventListener("keydown", handler);
     return () => document.removeEventListener("keydown", handler);
-  }, [callback]);
+  }, [callbackHandler]);
 };
